Refactor Wobbly schema to support Pydantic
Instead of writing an explicit object conversion function, rework
the Wobbly schema so that ORM objects can be converted directly to
Pydantic objects using `model_validate`. This requires moving the
ordering of results into the ORM rather than doing it manually,
moving the error attributes into a table and allowing more than one
error, and reworking the schema for results to avoid an autogenerated
key that clashes with the client-provided result ID.

UWS does not allow multiple errors in the current schema, but any
revision of the standard should ideally support multiple errors.
For now, we can follow the convention of storing at most one error
and not use this schema functionality, but since it makes the
Pydantic integration easier, it seemed better to future-proof now.

Change the cascade configuration of the job ORM model to ensure that
any orphaned result or error objects are deleted, which was already
required by the schema.