Add pagination support and more admin routes
Convert the `/jobs` route to support pagination. Change `count` to
`limit` and `after` to `since` to match Gafaelfawr and be more
precise (`after` could be interpreted as a generic sorting term,
while `since` is always about time), since we're not following the
current IVOA UWS API anyway.

Convert the existing admin jobs route to pagination and add multiple
new admin routes, allowing the listing of users and jobs for all
services and, in the later case, for all users as well.

Refactor the additional limiting criteria for searches into a
dependency so that it can be shared by the various routes.