Use models from Safir
Now that Safir has integrated Wobbly support, use the Pydantic
models from Safir for the API. If Safir changes in the future, we
may need to copy the older modules to support multiple versions of
the API, but for the time being this is simpler.

Change the `parameters` attribute to `json_parameters` following
the Safir model change, which allows Safir to use `parameters` for
the deserialized parameters that have been converted to the service
Pydantic model. Wobbly will never have access to that model, so it
only uses `json_parameters` and `SerializedJob` internally.